修必无向图、有向图的字段类型

+++ src/graph/breadth_first_paths.rs
@@ -5,14 +5,14 @@ use std::collections::VecDeque;
 use super::Graph;
 
 pub struct BreadthFirstPaths {
-    marked: Vec<bool>, // 是否已访问过。
-    edge_to: Vec<i32>, // 到顶点的路径上最后一个顶点。
-    s: i32,            // 起点。
+    marked: Vec<bool>,   // 是否已访问过。
+    edge_to: Vec<usize>, // 到顶点的路径上最后一个顶点。
+    s: usize,            // 起点。
 }
 
 impl BreadthFirstPaths {
     /// 创建图的广度优先结果。
-    pub fn new(g: Graph, s: i32) -> Self {
+    pub fn new(g: Graph, s: usize) -> Self {
         let marked = vec![false; g.v()];
         let edge_to = vec![0; g.v()];
         let mut d = Self { marked, edge_to, s };
@@ -21,16 +21,16 @@ impl BreadthFirstPaths {
     }
 
     /// 广度优先算法。
-    fn bfs(&mut self, g: &Graph, s: i32) {
+    fn bfs(&mut self, g: &Graph, s: usize) {
         let mut queue = VecDeque::new();
         queue.push_back(s);
 
-        self.marked[s as usize] = true;
+        self.marked[s] = true;
         while let Some(value) = queue.pop_front() {
             for w in g.adj(value) {
-                if !self.marked[*w as usize] {
-                    self.marked[*w as usize] = true; // 找到最短路径。
-                    self.edge_to[*w as usize] = value; // 最短路径的最后一条边。
+                if !self.marked[*w] {
+                    self.marked[*w] = true; // 找到最短路径。
+                    self.edge_to[*w] = value; // 最短路径的最后一条边。
                     queue.push_back(*w); // 添加最短路径。
                 }
             }
@@ -38,12 +38,12 @@ impl BreadthFirstPaths {
     }
 
     /// 是否存在 s 到 v 的路径。
-    pub fn has_path_to(&self, v: i32) -> bool {
-        self.marked[v as usize]
+    pub fn has_path_to(&self, v: usize) -> bool {
+        self.marked[v]
     }
 
     /// 获取 s 到 v 的路径。
-    pub fn path_to(&self, v: i32) -> Vec<i32> {
+    pub fn path_to(&self, v: usize) -> Vec<usize> {
         if !self.has_path_to(v) {
             return vec![];
         }
@@ -52,7 +52,7 @@ impl BreadthFirstPaths {
         let mut x = v;
         while x != self.s {
             path.push(x);
-            x = self.edge_to[x as usize];
+            x = self.edge_to[x];
         }
         path.push(self.s);
         path
@@ -61,8 +61,8 @@ impl BreadthFirstPaths {
 
 #[cfg(test)]
 mod tests {
-    use crate::graph::graph_data;
     use super::*;
+    use crate::graph::graph_data;
 
     #[test]
     fn has_path_to_test() {
